功能更新
1.添加路西洪哥拓哥阿鸡固定标识头像，登录后自动读取
2.优化了登录状态保持
3.添加登录密码校验及提示
4.氪金登记页面氪金时间会自动读取当前时间
5.氪金登记页面会根据当前登录的用户自动选择相应氪金人
6.进入氪金登记页面时如果未登录，会自动跳转至登录页

+++ yinjian-showtime/src/assets/js/pages/auth.init.js
@@ -26,6 +26,17 @@ function userLogin() {
         contentType: 'application/json;charset=utf-8',
         success: function(result) {
             console.log(result);
+            if (result.error == 'error') {
+                Swal.fire(
+                    {
+                        title: "脑瘫啊，没上去！",
+                        text: '用户名和密码也不对啊，你来这干得儿来了???',
+                        icon: 'error',
+                        confirmButtonColor: '#5438dc'
+                    }
+                )
+                return;
+            }
             localStorage.setItem("token", result.token);
             window.location.href = "/index.html"
         },
